Refactor

+++ Hana Engine_v.1/Hana Engine/Source/ModuleWindow.cpp
@@ -27,8 +27,8 @@ bool ModuleWindow::Init()
 		//Create window
 		SDL_DisplayMode DM;
 		SDL_GetCurrentDisplayMode(0, &DM);
-		w_width = DM.w*0.5;
-		w_height = DM.h*0.5;
+		w_width = int(DM.w * 0.5);
+		w_height = int(DM.h * 0.5);
 		Uint32 flags = SDL_WINDOW_SHOWN | SDL_WINDOW_OPENGL | SDL_WINDOW_RESIZABLE | SDL_WINDOW_MAXIMIZED;
 
 		if(FULLSCREEN == true)
